requirements.txt
fastapi
uvicorn
jinja2
python-multipart
pymongo
python-dotenv
langchain
langchain-groq
langchain-community
langchain-huggingface
langgraph
faiss-cpu
beautifulsoup4
pydantic
requests
bcrypt
pyjwt
langgraph-checkpoint-mongodb
langchain_huggingface
sentence-transformers
gitignore